+++ src/gocept/country/tests.py
@@ -1,16 +1,15 @@
 # vim:fileencoding=utf-8
 import doctest
-import os.path
-import zope.app.testing.functional
+import gocept.country
+import zope.app.wsgi.testlayer
 import zope.testing
 
-layer = zope.app.testing.functional.ZCMLLayer(
-    os.path.join(os.path.dirname(__file__), 'ftesting.zcml'),
-    __name__, 'gocept.country Layer', allow_teardown=True)
+layer = zope.app.wsgi.testlayer.BrowserLayer(
+    gocept.country, allowTearDown=True)
 
 
 def test_suite():
-    suite = zope.app.testing.functional.FunctionalDocFileSuite(
+    suite = doctest.DocFileSuite(
         'README.txt', optionflags=doctest.ELLIPSIS)
     suite.layer = layer
     return suite